move auth middleware into own auth folder, refactor code
