+++ src/pages/api/pantry/index.js
@@ -1,5 +1,5 @@
 const { Pantry } = require('../../../db/model/index.js')
-import handleDecodeJWT from "@/utils/handleDecodeJWT.js";
+import handleDecodeJWT from "@/utils/auth/handleDecodeJWT.js";
 
 // Pantry CRUD operation methods
 export default async function handler(req, res) {
