+++ src/pages/feed.jsx
@@ -1,5 +1,3 @@
-import { useState } from 'react'
-import FeedComponent from "@/features/feed/feed"
 import PageHeading from '@/features/feed/headings/pageHeading'
 import PostContainer from '@/features/feed/postContainer'
 
@@ -9,9 +7,6 @@ export default function FeedPage() {
             <div className="max-w-[1280px] w-full h-full px-2 md:px-4 py-6 ">
                 <PageHeading />
                 <PostContainer />
-
-                <section className='mt-24'>
-                </section>
             </div>
         </div>
