+++ src/pages/index.jsx
@@ -1,4 +1,5 @@
 import Image from 'next/image'
+import { motion as m, AnimatePresence } from 'framer-motion'
 import pantry from '../../public/images/pantry.png'
 import kitchen from '../../public/images/kitchen.png'
 import {
@@ -14,7 +15,7 @@ import {
   CalendarDaysIcon,
   UsersIcon,
 } from '@heroicons/react/24/outline'
-import { motion as m, AnimatePresence } from 'framer-motion'
+
 
 export default function LandingPage() {
   return (
