+++ src/pages/signup.jsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react'
 import { AiFillEyeInvisible, AiFillEye } from 'react-icons/ai'
-import authService from '@/utils/authService'
+import authService from '@/utils/auth/authService'
 
 function SignUp() {
   useEffect(() => {
@@ -8,7 +8,7 @@ function SignUp() {
       window.location.assign('/pantry')
     }
   })
-  
+
   const [signUpForm, setSignUpForm] = useState({
     username: '',
     email: '',
@@ -78,9 +78,9 @@ function SignUp() {
   }
 
   return (
-    <div className='flex min-h-full flex-col justify-center px-6 py-12 lg:px-8'>
+    <div className='flex h-full md:h-screen flex-col my-12 mb-24 md:mb-0 md:mt-36 px-6  lg:px-8'>
       <div className='sm:mx-auto sm:w-full sm:max-w-sm'>
-        <h2 className='mt-10 text-center text-2xl font-bold leading-9 tracking-tight text-gray-900'>
+        <h2 className=' text-center text-2xl font-bold leading-9 tracking-tight text-gray-900'>
           Create an account
         </h2>
       </div>
